Debugger has been improved more. It stores the most recent instructions, a bug with the stack trace maintanence has been fixed, and it now allows the user to read/write memory locations on the fly using the menu.
A hack has been found that will allow Super Mario Land to be played. When the PC = 3:0x67AA, the IF register (0xFF0F) has to be set to 0xE0 (or any other value where bit 4 is cleared). I'm investigating the root cause. It probably has something to do with the interrupt system.